# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"","version":"0.0.1"},"host":"localhost:80","basePath":"/focus","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/tasks":{"get":{"tags":["task"],"summary":"list task","description":"List all tasks.","operationId":"task#list","parameters":[{"name":"parent_id","in":"query","description":"The ID of the parent task","required":false,"type":"string"},{"name":"recursive","in":"query","description":"Whether to include all subtasks recursively","required":false,"type":"boolean"},{"name":"authorization","in":"header","description":"The authorization header","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/TaskTaskdetailResponseCollection"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/TaskListUnauthorizedResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/TaskListInternalServerErrorResponseBody"}}},"schemes":["http"]},"post":{"tags":["task"],"summary":"create task","description":"Create a new task.","operationId":"task#create","parameters":[{"name":"authorization","in":"header","description":"The authorization header","required":true,"type":"string"},{"name":"CreateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/CreateTaskInput","required":["title"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/Createtaskoutput"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/TaskCreateUnauthorizedResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/TaskCreateInternalServerErrorResponseBody"}}},"schemes":["http"]}},"/tasks/setup":{"post":{"tags":["task"],"summary":"setup task","description":"Setup the task service.","operationId":"task#setup","parameters":[{"name":"authorization","in":"header","description":"The authorization header","required":true,"type":"string"}],"responses":{"200":{"description":"OK response."},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/TaskSetupUnauthorizedResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/TaskSetupInternalServerErrorResponseBody"}}},"schemes":["http"]}},"/tasks/{task_id}":{"delete":{"tags":["task"],"summary":"delete task","description":"Delete a task.","operationId":"task#delete","parameters":[{"name":"task_id","in":"path","description":"The ID of the task","required":true,"type":"string"},{"name":"authorization","in":"header","description":"The authorization header","required":true,"type":"string"}],"responses":{"204":{"description":"No Content response."},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/TaskDeleteUnauthorizedResponseBody"}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/TaskDeleteTaskNotFoundResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/TaskDeleteInternalServerErrorResponseBody"}}},"schemes":["http"]},"patch":{"tags":["task"],"summary":"update task","description":"Update a task.","operationId":"task#update","parameters":[{"name":"task_id","in":"path","description":"The ID of the task","required":true,"type":"string"},{"name":"authorization","in":"header","description":"The authorization header","required":true,"type":"string"},{"name":"UpdateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/TaskUpdateInput"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/Taskdetail"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/TaskUpdateUnauthorizedResponseBody"}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/TaskUpdateTaskNotFoundResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/TaskUpdateInternalServerErrorResponseBody"}}},"schemes":["http"]}}},"definitions":{"CreateTaskInput":{"title":"CreateTaskInput","type":"object","properties":{"parent_id":{"type":"string","description":"The parent ID of the task","example":"Distinctio molestiae harum aut enim temporibus voluptas."},"title":{"type":"string","description":"The title of the task","example":"Praesentium beatae harum expedita consequuntur."}},"example":{"parent_id":"Et quasi.","title":"Et eligendi accusamus at accusantium qui quod."},"required":["title"]},"Createtaskoutput":{"title":"Mediatype identifier: createtaskoutput; view=default","type":"object","properties":{"created_at":{"type":"integer","description":"The timestamp when the task was created","example":4935846454160954655,"format":"int64"},"id":{"type":"string","description":"The ID of the task","example":"Fugit consequatur quas nobis maiores in officiis."}},"description":"CreateResponseBody result type (default view)","example":{"created_at":2443785852375484901,"id":"Impedit rem autem assumenda fuga et."},"required":["id","created_at"]},"TaskCreateInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Internal server error (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"TaskCreateUnauthorizedResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Unauthorized (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TaskDeleteInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Internal server error (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"TaskDeleteTaskNotFoundResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Task not found (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"TaskDeleteUnauthorizedResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Unauthorized (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TaskListInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Internal server error (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TaskListUnauthorizedResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Unauthorized (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TaskSetupInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Internal server error (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TaskSetupUnauthorizedResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Unauthorized (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TaskTaskdetailResponseCollection":{"title":"Mediatype identifier: taskdetail; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/TaskdetailResponse"},"description":"ListResponseBody is the result type for an array of TaskdetailResponse (default view)","example":[{"actual_time":3197877266714545109,"completed_at":5230743633861895822,"created_at":6868630231309701643,"estimated_time":6688229350677013194,"id":"Sint aut suscipit fugit aspernatur eos.","is_leaf":true,"lead_time":7964136735715035891,"parent_id":"Vel quas dignissimos reiciendis voluptatem tempore ut.","started_at":7800690728553666095,"status":"Sunt facilis quia.","title":"Voluptatum unde voluptatem laborum odio."},{"actual_time":3197877266714545109,"completed_at":5230743633861895822,"created_at":6868630231309701643,"estimated_time":6688229350677013194,"id":"Sint aut suscipit fugit aspernatur eos.","is_leaf":true,"lead_time":7964136735715035891,"parent_id":"Vel quas dignissimos reiciendis voluptatem tempore ut.","started_at":7800690728553666095,"status":"Sunt facilis quia.","title":"Voluptatum unde voluptatem laborum odio."}]},"TaskUpdateInput":{"title":"TaskUpdateInput","type":"object","properties":{"estimated_time":{"type":"integer","description":"The estimated time of the task","example":1990979614394948437,"format":"int64"},"next_id":{"type":"string","description":"The next ID of the task","example":"Repellendus excepturi possimus."},"parent_id":{"type":"string","description":"The parent ID of the task","example":"Eos dicta earum nihil."},"status":{"type":"string","description":"The status of the task","example":"Cum dicta."},"title":{"type":"string","description":"The title of the task","example":"Optio velit."}},"example":{"estimated_time":1850693608130154120,"next_id":"Aut deleniti a accusamus deserunt odit.","parent_id":"Dignissimos rerum corrupti.","status":"Quis eum.","title":"Et quo."}},"TaskUpdateInternalServerErrorResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Internal server error (default view)","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"TaskUpdateTaskNotFoundResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Task not found (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"TaskUpdateUnauthorizedResponseBody":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"description":"Unauthorized (default view)","example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"Taskdetail":{"title":"Mediatype identifier: taskdetail; view=default","type":"object","properties":{"actual_time":{"type":"integer","description":"The actual time of the task","example":8089272436431874333,"format":"int64"},"completed_at":{"type":"integer","description":"The timestamp when the task was completed","example":8894326357489074681,"format":"int64"},"created_at":{"type":"integer","description":"The timestamp when the task was created","example":35803148126737651,"format":"int64"},"estimated_time":{"type":"integer","description":"The estimated time of the task","example":4869562914718477248,"format":"int64"},"id":{"type":"string","description":"The ID of the task","example":"Aliquid non."},"is_leaf":{"type":"boolean","description":"Whether the task is a leaf task","example":true},"lead_time":{"type":"integer","description":"The lead time of the task","example":6064961824949202856,"format":"int64"},"parent_id":{"type":"string","description":"The parent ID of the task","example":"Alias est nulla eveniet earum."},"started_at":{"type":"integer","description":"The timestamp when the task was started","example":8281130147352953431,"format":"int64"},"status":{"type":"string","description":"The status of the task","example":"Explicabo sed distinctio."},"title":{"type":"string","description":"The title of the task","example":"Nulla ab pariatur."}},"description":"UpdateResponseBody result type (default view)","example":{"actual_time":7748456940911302561,"completed_at":610255816507690574,"created_at":6947351725023923149,"estimated_time":3018525635946327686,"id":"Placeat voluptatem mollitia.","is_leaf":true,"lead_time":6160651272351825989,"parent_id":"Omnis perspiciatis.","started_at":7109543597677480911,"status":"Perspiciatis perferendis nemo.","title":"Suscipit harum impedit laboriosam dolor dignissimos."},"required":["id","title","created_at","status"]},"TaskdetailResponse":{"title":"Mediatype identifier: taskdetail; view=default","type":"object","properties":{"actual_time":{"type":"integer","description":"The actual time of the task","example":3038736976794975442,"format":"int64"},"completed_at":{"type":"integer","description":"The timestamp when the task was completed","example":3707321435505658767,"format":"int64"},"created_at":{"type":"integer","description":"The timestamp when the task was created","example":7851324801854454987,"format":"int64"},"estimated_time":{"type":"integer","description":"The estimated time of the task","example":4280074667404497147,"format":"int64"},"id":{"type":"string","description":"The ID of the task","example":"Porro unde unde."},"is_leaf":{"type":"boolean","description":"Whether the task is a leaf task","example":false},"lead_time":{"type":"integer","description":"The lead time of the task","example":5645038487346972198,"format":"int64"},"parent_id":{"type":"string","description":"The parent ID of the task","example":"Officia quas ullam culpa dolorem."},"started_at":{"type":"integer","description":"The timestamp when the task was started","example":1140812738329248349,"format":"int64"},"status":{"type":"string","description":"The status of the task","example":"Cum blanditiis."},"title":{"type":"string","description":"The title of the task","example":"Magnam laboriosam quam."}},"description":"TaskdetailResponse result type (default view)","example":{"actual_time":3435329221889764974,"completed_at":8592677132735821252,"created_at":3531551450490652910,"estimated_time":5352798065806699126,"id":"Recusandae et et qui.","is_leaf":false,"lead_time":886280517307206378,"parent_id":"Veritatis quos.","started_at":867276528224863209,"status":"Sapiente voluptas.","title":"Et nisi quae dolores."},"required":["id","title","created_at","status"]}}}